faceapi/requirements.txt
Flask
requests
Pillow
flask-cors
transformers
openai
python-dotenv
keybert
kiwipiepy
google-genai
mediapipe
opencv-python-headless
numpy
langgraph
langchain
langchain-community
langchain-openai
chromadb
sentence-transformers
beautifulsoup4
